src/requirements.txt
Flask==1.0
mysql-connector-python